Address nenad1002 API review feedback from PR #500
- Remove pub use Result re-export from lib.rs to avoid shadowing std::Result
- Remove ModelLoadManager from public API
- Add builder pattern for FoundryLocalConfig (private fields, chainable setters)
- Add Logger trait stub with builder method
- Change start_web_service to return Result<()>; use urls() separately
- Change get_loaded_models to return Vec<Arc<ModelVariant>> instead of Vec<String>
- Use interior mutability (AtomicUsize) for select_variant(&self)
- Implement custom Debug for Model/ModelVariant (no internal field leakage)
- Replace serde_json::Value with typed TranscriptionSegment/TranscriptionWord structs
- Change ModelInfo integer fields to u64 (version, file_size_mb, max_output_tokens, created_at_unix)
- Update docs/api.md to reflect all API changes
- Fix foundry-local-webserver sample for new start_web_service signature

+++ sdk_v2/rust/src/types.rs
@@ -55,7 +55,7 @@ pub struct ModelSettings {
 pub struct ModelInfo {
     pub id: String,
     pub name: String,
-    pub version: i64,
+    pub version: u64,
     pub alias: String,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub display_name: Option<String>,
@@ -78,15 +78,15 @@ pub struct ModelInfo {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub runtime: Option<Runtime>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub file_size_mb: Option<f64>,
+    pub file_size_mb: Option<u64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub supports_tool_calling: Option<bool>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_output_tokens: Option<i64>,
+    pub max_output_tokens: Option<u64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub min_fl_version: Option<String>,
     #[serde(default)]
-    pub created_at_unix: i64,
+    pub created_at_unix: u64,
 }
 
 /// Desired response format for chat completions.
